Refactor code to decouple source code from test code
Some test util functions where moved to logprep source util. This
prevents the need of development libraries such as 'pytest' to be
imported in production. Previously when logprep was called with the
argument --validate-rules test code was imported, this is not the case
anymore.
