+++ logprep/util/rule_dry_runner.py
@@ -12,18 +12,49 @@ import json
 from ruamel.yaml import YAML
 from colorama import Fore, Back
 
-from logprep.util.helper import print_color
-
-from tests.acceptance.util import (create_temporary_config_file_at_path, get_test_output_multi,
-                                   parse_json, parse_jsonl, recursive_compare)
+from logprep.runner import Runner
+from logprep.util.helper import print_color, recursive_compare, remove_file_if_exists
+from logprep.util.json_handling import dump_config_as_file, parse_jsonl, parse_json
 
 yaml = YAML(typ='safe', pure=True)
+
+
+def get_test_output_multi(config_path, logger):
+    patched_runner = get_patched_runner(config_path, logger)
+
+    parsed_outputs = list()
+    output_paths = list()
+    output_paths.append(patched_runner._configuration['connector'].get('output_path', None))
+    output_paths.append(patched_runner._configuration['connector'].get('output_path_custom', None))
+    output_paths.append(patched_runner._configuration['connector'].get('output_path_errors', None))
+    output_paths = [output_path for output_path in output_paths if output_path]
+
+    for output_path in output_paths:
+        remove_file_if_exists(output_path)
+
+    patched_runner.start()
+
+    for output_path in output_paths:
+        parsed_outputs.append(parse_jsonl(output_path))
+        remove_file_if_exists(output_path)
+
+    return parsed_outputs
+
+
+def get_patched_runner(config_path, logger):
+    runner = Runner(bypass_check_to_obtain_non_singleton_instance=True)
+    runner.set_logger(logger)
+    runner.load_configuration(config_path)
+    # patch runner to stop on empty pipeline
+    runner._keep_iterating = lambda: False
+
+    return runner
 
 
 class DryRunner:
     """Used to run pipeline with given events and show changes made by processing."""
 
-    def __init__(self, dry_run: str, config_path: str, full_output: str, use_json: bool):
+    def __init__(self, dry_run: str, config_path: str, full_output: str, use_json: bool, logger):
         with open(config_path, 'r') as yaml_file:
             self._config_yml = yaml.load(yaml_file)
 
@@ -31,10 +62,12 @@ class DryRunner:
         self._use_json = use_json
 
         self._config_yml['connector'] = {
-                'type': 'writer_json_input' if use_json else 'writer',
-                'input_path': dry_run
-            }
+            'type': 'writer_json_input' if use_json else 'writer',
+            'input_path': dry_run
+        }
         self._config_yml['process_count'] = 1
+
+        self._logger = logger
 
     def run(self):
         """Run the dry runner."""
@@ -47,9 +80,9 @@ class DryRunner:
                                                                         'output_errors.jsonl')
 
         config_path = path.join(tmp_path, 'generated_config.yml')
-        create_temporary_config_file_at_path(config_path, self._config_yml)
+        dump_config_as_file(config_path, self._config_yml)
 
-        test_output, test_output_custom, test_output_error = get_test_output_multi(config_path)
+        test_output, test_output_custom, test_output_error = get_test_output_multi(config_path, self._logger)
 
         input_path = self._config_yml['connector']['input_path']
         input_data = parse_json(input_path) if self._use_json else parse_jsonl(input_path)
